third_party/WebKit/LayoutTests/fast/lists/list-and-margin-collapse.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<ul id="target" style="margin-top:100px;">
  <li>
    <div style="overflow:hidden; margin-top:100px; height:25px;"><a href="#">xxx</a></div>
  </li>
</ul>
<div id="log"></div>
<script>
test(function() {
  var height = document.getElementById("target").clientHeight;
  assert_equals(height, 25, "the height of ul should be 25px")
}, "list and margin collapse");
</script>
